+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fix mouse wheel scrolling when viewing diff (<https://github.com/lapce/lapce/issues/3821>)
 - Fix editor tabs not selectable while appearing selectable
 - Fix flickering when reordering editor tabs
+- (Un)confirmed state of the editor is properly (re)stored
 
 ## 0.4.5